Reordered Folders for Project
Also added gradient and header button image style
