+++ iRAK/Assets.xcassets/Colors/Foreground2Color.colorset/Contents.json
