+++ iRAK/Assets.xcassets/Colors/ForegroundColor.colorset/Contents.json
