+++ iRAK/Assets.xcassets/SvgIcons/x-symbol.imageset/Contents.json
